requirements.txt
google-genai
pydantic
streamlit
python-dotenv
Pillow